Remove non-sh test(1)
Some shells -- such as zsh, GNU Bash, ksh, and more -- have a built-in
test(1)-like syntax called `[[`. Since it's a shell built-in, it does
not incur the extra cost of booting a process.

However, it is not part of POSIX sh and therefore not available under
`/bin/sh` on Debian.

Switch to the `[` (aka test(1)) command instead.